fix[project]: изменена структура проекта
