--- Controller.cs
@@ -1,10 +1,0 @@
-namespace PromAss
-{
-    /// <summary>
-    /// Основной класс 
-    /// </summary>
-    class Controller
-    {
-
-    }
-}
