--- dao/SerialConnection.cs
@@ -1,30 +1,0 @@
-using System;
-using System.IO.Ports;
-
-namespace PromAss
-{
-    class SerialConnection : IConnectable
-    {
-        private SerialPort port;
-
-        public void Connect()
-        {
-
-        }
-
-        public void Disconnect()
-        {
-
-        }
-
-        public object ReadData()
-        {
-            return null;
-        }
-
-        public void SendMessage(object message)
-        {
-
-        }
-    }
-}
