--- model/Frame.cs
@@ -1,11 +1,0 @@
-namespace PromAss
-{
-    /// <summary>
-    /// Считанные данные с датчиков преобразуются 
-    /// в Frame для дальнейшей передачи в Прометей
-    /// </summary>
-    public class Frame
-    {
-
-    }
-}
