+++ src/Controller.cs
@@ -1,0 +1,7 @@
+/// <summary>
+/// Основной класс 
+/// </summary>
+class Controller
+{
+
+}
