+++ src/dao/IConnectable.cs
@@ -8,12 +8,12 @@ interface IConnectable
     /// <summary>
     /// Осуществляет подключение объекта к устройству
     /// </summary>
-    void Connect();
+    void Connect(); 
 
     /// <summary>
     /// Осуществляет отключение 
     /// </summary>
-    void Disconnect();
+    void Disconnect(); 
 
     /// <summary>
     /// Прием данных
