+++ src/dao/SerialConnection.cs
@@ -1,0 +1,27 @@
+using System;
+using System.IO.Ports;
+
+public class SerialConnection : IConnectable
+{
+    private SerialPort port;
+
+    public void Connect()
+    {
+
+    }
+
+    public void Disconnect()
+    {
+
+    }
+
+    public object ReadData()
+    {
+        return null;
+    }
+
+    public void SendMessage(object message)
+    {
+
+    }
+}
