+++ src/model/Frame.cs
@@ -1,0 +1,8 @@
+/// <summary>
+/// Считанные данные с датчиков преобразуются 
+/// в Frame для дальнейшей передачи в Прометей
+/// </summary>
+public class Frame
+{
+
+}
